+++ docs/changes.rst
@@ -38,6 +38,11 @@ Release history
 
         * :attr:`HGCallPolicies.business_continuity <wxc_sdk.telephony.huntgroup.HGCallPolicies.business_continuity>`
 - feat: location PSTN settings API: :attr:`api.telephony.pstn <wxc_sdk.telephony.TelephonyApi.pstn>`
+- feat: new API, organisation level emergency settings :attr:`api.telephony.emergency_services <wxc_sdk.telephony.TelephonyApi.emergency_services>`
+- feat: new API, location level emergency settings :attr:`api.telephony.location.emergency_services <wxc_sdk.telephony.location.TelephonyLocationApi.emergency_services>`
+- feat: new API, user ECBN settings :attr:`api.person_settings.ecbn <wxc_sdk.person_settings.PersonSettingsApi.ecbn>`
+- feat: new API, virtual line ECBN settings :attr:`api.telephony.virtual_lines.ecbn <wxc_sdk.telephony.virtual_line.VirtualLinesApi.ecbn>`
+- feat: new API, workspace ECBN settings :attr:`api.workspace_settings.ecbn <wxc_sdk.workspace_settings.WorkspaceSettingsApi.ecbn>`
 
 1.21.1
 ------